@@ -17,47 +17,47 @@ class BraintreeDropInRequest {
 
   /// Authorization allowing this client to communicate with Braintree.
   /// Either [clientToken] or [tokenizationKey] must be set.
-  String clientToken;
+  String? clientToken;
 
   /// Authorization allowing this client to communicate with Braintree.
   /// Either [clientToken] or [tokenizationKey] must be set.
-  String tokenizationKey;
+  String? tokenizationKey;
 
   /// Amount for the transaction. This is only used for 3D secure verfications.
-  String amount;
+  String? amount;
 
   /// Whether the Drop-in should collect and return device data for fraud prevention.
-  bool collectDeviceData;
+  bool? collectDeviceData;
 
   /// If 3D Secure has been enabled in the control panel and an amount is specified in
   /// [amount], Drop-In will request a 3D Secure verification for any new cards added by the user.
-  bool requestThreeDSecureVerification;
+  bool? requestThreeDSecureVerification;
 
   /// Google Payment request. Google Pay will be disabled if this is set to `null`.
-  BraintreeGooglePaymentRequest googlePaymentRequest;
+  BraintreeGooglePaymentRequest? googlePaymentRequest;
 
   /// PayPal request. PayPal will be disabled if this is set to `null`.
-  BraintreePayPalRequest paypalRequest;
+  BraintreePayPalRequest? paypalRequest;
 
   /// Whether Venmo should be enabled.
-  bool venmoEnabled;
+  bool? venmoEnabled;
 
   /// Whether cards should be enabled.
-  bool cardEnabled;
+  bool? cardEnabled;
 
   /// Whether the card number should be masked if the field is not focused.
-  bool maskCardNumber;
+  bool? maskCardNumber;
 
   /// Whether the security code should be masked during input.
-  bool maskSecurityCode;
+  bool? maskSecurityCode;
 
   /// Whether customers should be allowed to manage their vaulted payment methods.
-  bool vaultManagerEnabled;
+  bool? vaultManagerEnabled;
 
   /// ApplePay request. ApplePay will be disabled if this is set to `null`.
   /// The ApplePay option will not be visible in the drop-in UI if the setup in
   /// Xcode, App Store Connect or Braintree control panel was done incorrectly.
-  BraintreeApplePayRequest applePayRequest;
+  BraintreeApplePayRequest? applePayRequest;
 
   /// Converts this request object into a JSON-encodable format.
   Map<String, dynamic> toJson() => {
@@ -68,10 +68,10 @@ class BraintreeDropInRequest {
         if (requestThreeDSecureVerification != null)
           'requestThreeDSecureVerification': requestThreeDSecureVerification,
         if (googlePaymentRequest != null)
-          'googlePaymentRequest': googlePaymentRequest.toJson(),
-        if (paypalRequest != null) 'paypalRequest': paypalRequest.toJson(),
+          'googlePaymentRequest': googlePaymentRequest!.toJson(),
+        if (paypalRequest != null) 'paypalRequest': paypalRequest!.toJson(),
         if (applePayRequest != null)
-          'applePayRequest': applePayRequest.toJson(),
+          'applePayRequest': applePayRequest!.toJson(),
         if (venmoEnabled != null) 'venmoEnabled': venmoEnabled,
         if (cardEnabled != null) 'cardEnabled': cardEnabled,
         if (maskCardNumber != null) 'maskCardNumber': maskCardNumber,
@@ -89,13 +89,13 @@ class BraintreeCreditCardRequest {
   });
 
   /// Number shown on the credit card.
-  String cardNumber;
+  String? cardNumber;
 
   /// Two didgit expiration month, e.g. `'05'`.
-  String expirationMonth;
+  String? expirationMonth;
 
   /// Four didgit expiration year, e.g. `'2021'`.
-  String expirationYear;
+  String? expirationYear;
 
   Map<String, dynamic> toJson() => {
         if (cardNumber != null) 'cardNumber': cardNumber,
@@ -113,16 +113,16 @@ class BraintreeGooglePaymentRequest {
   });
 
   /// Total price of the payment.
-  String totalPrice;
+  String? totalPrice;
 
   /// Currency code of the transaction.
-  String currencyCode;
+  String? currencyCode;
 
   /// Whether billing address information should be collected and passed.
-  bool billingAddressRequired;
+  bool? billingAddressRequired;
 
   /// Google Merchant ID. Optional in sandbox, but if set, must be a valid production Google Merchant ID.
-  String googleMerchantID;
+  String? googleMerchantID;
 
   /// Converts this request object into a JSON-encodable format.
   Map<String, dynamic> toJson() => {
@@ -144,16 +144,16 @@ class BraintreePayPalRequest {
 
   /// Amount of the transaction. If [amount] is `null`, PayPal will use the billing agreement (Vault) flow.
   /// If [amount] is set, PayPal will follow the one time payment (Checkout) flow.
-  String amount;
+  String? amount;
 
   /// Currency code. If set to null`null`, PayPal will choose it based on the active merchant account in the client token.
-  String currencyCode;
+  String? currencyCode;
 
   /// The merchant name displayed in the PayPal flow. If set to `null`, PayPal will use the company name in your Braintree account.
-  String displayName;
+  String? displayName;
 
   /// Description for the billing agreement for the Vault flow.
-  String billingAgreementDescription;
+  String? billingAgreementDescription;
 
   /// Converts this request object into a JSON-encodable format.
   Map<String, dynamic> toJson() => {
@@ -167,30 +167,27 @@ class BraintreePayPalRequest {
 
 class BraintreeApplePayRequest {
   BraintreeApplePayRequest({
-    this.amount,
-    this.displayName,
-    this.currencyCode,
-    this.countryCode,
-    this.appleMerchantID,
-  })  : assert(amount != null),
-        assert(displayName != null),
-        assert(currencyCode != null && countryCode != null),
-        assert(appleMerchantID != null);
+    required this.amount,
+    required this.displayName,
+    required this.currencyCode,
+    required this.countryCode,
+    required this.appleMerchantID,
+  });
 
   /// The item's amount.
-  final double amount;
+  final double? amount;
 
   /// Short description of the item.
-  final String displayName;
+  final String? displayName;
 
   /// The three-letter ISO 4217 currency code.
-  final String currencyCode;
+  final String? currencyCode;
 
   /// The three-letter ISO 4217 currency code.
-  final String countryCode;
+  final String? countryCode;
 
   /// Apple merchant identifier.
-  final String appleMerchantID;
+  final String? appleMerchantID;
 
   /// Converts this request object into a JSON-encodable format.
   Map<String, dynamic> toJson() => {
