@@ -1,37 +1,48 @@
-import 'package:meta/meta.dart';
 
 class BraintreeDropInResult {
   const BraintreeDropInResult({
-    @required this.paymentMethodNonce,
-    @required this.deviceData,
+    required this.paymentMethodNonce,
+    required this.deviceData,
   });
 
-  factory BraintreeDropInResult.fromJson(dynamic source) {
-    if (source == null) return null;
+  factory BraintreeDropInResult.fromJson(dynamic? source) {
+    if(source == null){
+      return BraintreeDropInResult(
+        paymentMethodNonce: null,
+        deviceData: null,
+      );
+    }
+
     return BraintreeDropInResult(
-      paymentMethodNonce:
-          BraintreePaymentMethodNonce.fromJson(source['paymentMethodNonce']),
+      paymentMethodNonce: BraintreePaymentMethodNonce.fromJson(source['paymentMethodNonce']),
       deviceData: source['deviceData'],
     );
   }
 
   /// The payment method nonce containing all relevant information for the payment.
-  final BraintreePaymentMethodNonce paymentMethodNonce;
+  final BraintreePaymentMethodNonce? paymentMethodNonce;
 
   /// String of device data. `null`, if `collectDeviceData` was set to false.
-  final String deviceData;
+  final String? deviceData;
 }
 
 class BraintreePaymentMethodNonce {
   const BraintreePaymentMethodNonce({
-    @required this.nonce,
-    @required this.typeLabel,
-    @required this.description,
-    @required this.isDefault,
+    required this.nonce,
+    required this.typeLabel,
+    required this.description,
+    required this.isDefault,
   });
 
-  factory BraintreePaymentMethodNonce.fromJson(dynamic source) {
-    if (source == null) return null;
+  factory BraintreePaymentMethodNonce.fromJson(dynamic? source) {
+    if (source == null){
+      return BraintreePaymentMethodNonce(
+        nonce: null,
+        typeLabel: null,
+        description: null,
+        isDefault: null,
+      );
+    }
     return BraintreePaymentMethodNonce(
       nonce: source['nonce'],
       typeLabel: source['typeLabel'],
@@ -42,14 +53,14 @@ class BraintreePaymentMethodNonce {
 
   /// The nonce generated for this payment method by the Braintree gateway. The nonce will represent
   /// this PaymentMethod for the purposes of creating transactions and other monetary actions.
-  final String nonce;
+  final String? nonce;
 
   /// The type of this PaymentMethod for displaying to a customer, e.g. 'Visa'. Can be used for displaying appropriate logos, etc.
-  final String typeLabel;
+  final String? typeLabel;
 
   /// The description of this PaymentMethod for displaying to a customer, e.g. 'Visa ending in...'.
-  final String description;
+  final String? description;
 
   /// True if this payment method is the default for the current customer, false otherwise.
-  final bool isDefault;
+  final bool? isDefault;
 }
